wip
git-svn-id: https://www.dkfz.de/svn/B040/FCSTOOLS/trunk/QuickFit3@3890 0d8be23c-4016-4ca5-a69c-3788035ecc6c

+++ plugins/imagestack/qfrdrimagestack_parserfunctions.cpp
@@ -213,7 +213,7 @@ void fStack_getframe(qfmpResult &res, const qfmpResult *params, unsigned int n, 
 
         if (n>=3 && params[2].type==qfmpDouble) {
             frame=params[2].toInteger();
-            if (frame<0 || frame>=rdr->getImageStackFrames(stack)) {
+            if (frame<0 || frame>=(int)rdr->getImageStackFrames(stack)) {
                 parser->qfmpError(QObject::tr("imagestack_getframe(rdrid,stack,frame,...): the specified frame is not available in this RDR"));
                 res.setInvalid();
                 return;
